Cart: Change type of vatPercentage from Double to Float
Not even France will need a full Double for specifying VAT percentages

+++ izettle-cart/src/main/java/com/izettle/cart/CartUtils.java
@@ -193,10 +193,10 @@ class CartUtils {
 		return retList;
 	}
 
-	static Long calculateVatFromGrossAmount(final Long amountIncVat, final Double vatPercent) {
+	static Long calculateVatFromGrossAmount(final Long amountIncVat, final Float vatPercent) {
 		if (amountIncVat == null || vatPercent == null) {
 			return null;
 		}
-		return amountIncVat - round((amountIncVat * 100) / (100 + vatPercent));
+		return amountIncVat - round((amountIncVat * 100) / (100 + (double) vatPercent));
 	}
 }
